FIX : BDBD-333 postWritingData 함수 수정
